requirements.txt
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comtypes==1.2.0
cryptography==41.0.7
dotenv==0.0.5
grpcio==1.60.0
grpcio-tools==1.60.0
idna==3.6
numpy==1.26.2
protobuf==4.25.1
PyAudio==0.2.14
pycparser==2.21
pyOpenSSL==23.3.0
pypiwin32==223
python-dotenv==1.0.0
pyttsx3==2.90
pywin32==306
requests==2.31.0
sounddevice==0.4.6
soundfile==0.12.1
SpeechRecognition==3.10.1
srt==3.5.3
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
vosk==0.3.45
websockets==12.0
